Refactor: extract regex matchers for tests, relocate and update function visibility
